--- a/applications/nydus/Kubefile
+++ b/applications/nydus/Kubefile
@@ -1,0 +1,3 @@
+FROM kubernetes:v1.19.8
+COPY nydusdfile .
+COPY Metadata .