--- a/applications/nydus/nydusdfile/nydusd_scp_file/clean.sh
+++ b/applications/nydus/nydusdfile/nydusd_scp_file/clean.sh
@@ -1,0 +1,4 @@
+systemctl stop nydusd.service
+rm /etc/systemd/system/nydusd.service
+rm /usr/bin/nydusd
+rm /var/lib/sealer/nydusd.sh