--- a/applications/nydus/nydusdfile/nydusd_scp_file/nydusd.service
+++ b/applications/nydus/nydusdfile/nydusd_scp_file/nydusd.service
@@ -1,0 +1,8 @@
+[Unit]
+Description=sealer nydusd rootfs
+[Service]
+TimeoutStartSec=3
+ExecStart=/var/lib/sealer/nydusd.sh
+Restart=always
+[Install]
+WantedBy=multi-user.target