--- a/applications/nydus/nydusdfile/serverclean.sh
+++ b/applications/nydus/nydusdfile/serverclean.sh
@@ -1,0 +1,2 @@
+systemctl stop nydusd_http_server.service
+rm /etc/systemd/system/nydusd_http_server.service