<!--  Thanks for submitting a pull request! Here are some tips for you:
1. Please make sure you have read and understood the contributing guidelines: https://github.com/alibaba/blob/master/CONTRIBUTING.md
2. Please make sure the PR has a corresponding issue.
-->

### Describe what this PR does / why we need it
Sealer integration nydus to improve image distribution performance.
#### Performance comparison
| rootfs size      | inodes number | sealer-0.6.1 init time | sealer-nydusd init time              |
|------------------|---------------|------------------------|--------------------------------------|
| 770M (299 files) | 1             | 93s                    | 23s(include 6s image Conversion)     |
| 770M (299 files) | 6             | 95s                    | 28s(include 6s image Conversion)     |
| 30G (1299 files) | 1             | 16m40s                 | 6m20s(include 6min image Conversion) |
| 30G (1299 files) | 6             | 17m56s                 | 6m23s(include 6min image Conversion) |

The speed of image conversion is  limited by disk performance.

### Does this pull request fix one issue?

<!--If that, add "Fixes #xxxx" below in the next line. For example, Fixes #15. Otherwise, add "NONE" -->
Fixes #937
### Describe how you did it
If --nydusd = true and there are nydsudfile dir in rootfs, sealer will mount rootfs by nydus.