@fanux 2022-02-10
1. Move clean code to clean.sh
2. Add nydusFile system struct to implement filesyste.Interface