@nan1994 2022-02-16
test image: kubernetes-nydus:v1.19.8 
sealer run kubernetes-nydus:v1.19.8 --masters xxxxxxx  --passwd xxxxxx